requirements.txt
flask
flask_sqlalchemy
werkzeug
pyjwt